shared-py/pyproject.toml
[project]
name = "shared-py"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "nats-py>=2.9.0",
    "aiohttp>=3.11.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]

[tool.setuptools.package-dir]
shared_py = "src"
